pull out nested file type checkers
