+++ looper/models.py
@@ -63,7 +63,7 @@ import yaml as _yaml
 
 from . import LOOPERENV_VARNAME
 from exceptions import *
-from utils import partition
+from utils import bam_or_fastq, check_bam, check_fastq, partition
 
 COL_KEY_SUFFIX = "_key"
 ATTRDICT_METADATA = ("_force_nulls", "_attribute_identity")
@@ -1123,13 +1123,11 @@ class Sample(object):
         pieces together an actual path, by substituting variables (encoded by "{variable}"") with
         sample attributes.
 
-        :param column_name: Name of sample attribute (equivalently, sample sheet column) specifying a derived column.
-        :type column_name: str
-        :param source_key: The key of the data_source, used to index into the project config data_sources
+        :param str column_name: Name of sample attribute (equivalently, sample sheet column) specifying a derived column.
+        :param str source_key: The key of the data_source, used to index into the project config data_sources
         section. By default, the source key will be taken as the value of the specified column (as a sample
         attribute); but	for cases where the sample doesn't have this attribute yet (e.g. in a merge table),
         you must specify the source key.
-        :type source_key: str
         :param extra_vars: By default, locate_data_source will look to populate the template location
         using attributes found in the current sample; however, you may also provide a dict of
         extra variables that can also be used for variable replacement. These extra variables are
@@ -1177,6 +1175,7 @@ class Sample(object):
 
         return val
 
+
     def get_genome_transcriptome(self):
         """
         Get genome and transcriptome, based on project config file.
@@ -1187,12 +1186,12 @@ class Sample(object):
         except AttributeError:
             _LOGGER.warn("Project config lacks genome mapping for "
                               "organism '%s'", str(self.organism))
-        # get transcriptome
         try:
             self.transcriptome = getattr(self.prj.transcriptomes, self.organism)
         except AttributeError:
             _LOGGER.warn("Project config lacks transcriptome mapping for "
                               "organism '%s'", str(self.organism))
+
 
     def set_file_paths(self, override=False):
         """
@@ -1223,6 +1222,7 @@ class Sample(object):
         except:
             pass
 
+
     def make_sample_dirs(self):
         """
         Creates sample directory structure if it doesn't exist.
@@ -1230,6 +1230,7 @@ class Sample(object):
         for path in self.paths.__dict__.values():
             if not _os.path.exists(path):
                 _os.makedirs(path)
+
 
     def get_sheet_dict(self):
         """
@@ -1391,71 +1392,6 @@ class Sample(object):
                 missing_files.append(path)
             else:
                 existing_files.append(path)
-
-        import subprocess as sp
-        from collections import Counter
-
-        def bam_or_fastq(input_file):
-            """
-            Checks if string endswith `bam` or `fastq`.
-            Returns string. Raises TypeError if neither.
-
-            :param str input_file: String to check.
-            :return str: filetype (extension without dot prefix)
-            :raises TypeError: if file does not appear of a supported type,
-                based on extension
-            """
-            if input_file.endswith(".bam"):
-                return "bam"
-            elif input_file.endswith(".fastq") or \
-                    input_file.endswith(".fq") or \
-                    input_file.endswith(".fq.gz") or \
-                    input_file.endswith(".fastq.gz"):
-                return "fastq"
-            else:
-                raise TypeError("Type of input file ends in neither '.bam' "
-                                "nor '.fastq' [file: '" + input_file +"']")
-
-        def check_bam(bam, o):
-            """
-            Check reads in BAM file for read type and lengths.
-
-            :param bam: BAM file path.
-            :type bam: str
-            :param o: Number of reads to look at for estimation.
-            :type o: int
-            """
-            # view reads
-            try:
-                p = sp.Popen(['samtools', 'view', bam], stdout=sp.PIPE)
-
-                # Count paired alignments
-                paired = 0
-                read_length = Counter()
-                while o > 0:  # Count down number of lines
-                    line = p.stdout.next().split("\t")
-                    flag = int(line[1])
-                    read_length[len(line[9])] += 1
-                    if 1 & flag:  # check decimal flag contains 1 (paired)
-                        paired += 1
-                    o -= 1
-                p.kill()
-            except OSError:
-                reason = "Note (samtools not in path): For NGS inputs, " \
-                         "looper needs samtools to auto-populate " \
-                         "'read_length' and 'read_type' attributes; " \
-                         "these attributes were not populated."
-                raise OSError(reason)
-
-            _LOGGER.debug("Read lengths: {}".format(read_length))
-            _LOGGER.debug("paired: {}".format(paired))
-            return read_length, paired
-
-        def check_fastq(fastq, o):
-            """
-            """
-            raise NotImplementedError("Detection of read type/length for "
-                                      "fastq input is not yet implemented.")
 
         # for samples with multiple original bams, check all
         files = list()
